fix: get audio working when placing sound sources
This fix has many changes, the primary one being the replacement of the
many-to-many relationship between SoundEventAsset and SoundAsset with a
one-to-many relationship. It restructures the UI to support this. The
secondary change is delegating PhasePlayer to PhasePlayerFromUrl. In the
future, the delegation should be inlined and PhasePlayerFromUrl should
be removed.